+++ package.json
@@ -22,6 +22,7 @@
     "axios": "^0.17.1",
     "expo": "^22.0.2",
     "react": "16.0.0-beta.5",
-    "react-native": "^0.49.5"
+    "react-native": "^0.49.5",
+    "react-navigation": "^1.0.0-beta.20"
   }
 }
